requirements.txt
matplotlib
numpy
pandas
pyyaml
torch
vllm==0.6.2
transformers
accelerate